Resolve codacy warning and move tests to complex folder
