+++ jdi-light-bootstrap-tests/src/test/java/io/github/epam/bootstrap/tests/complex/MultipleInputsTests.java
@@ -1,4 +1,4 @@
-package io.github.epam.bootstrap.tests.common;
+package io.github.epam.bootstrap.tests.complex;
 
 import io.github.epam.TestsInit;
 import org.testng.annotations.BeforeClass;
@@ -9,13 +9,13 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.Map;
 
-import static com.epam.jdi.light.ui.bootstrap.elements.BootstrapUtils.isElementInViewPort;
+import static com.epam.jdi.light.ui.bootstrap.BootstrapUtils.isElementInViewPort;
 import static io.github.com.StaticSite.bsPage;
 import static io.github.com.pages.BootstrapPage.multipleInputs;
 import static io.github.com.pages.BootstrapPage.redButton;
 import static io.github.epam.bootstrap.tests.BaseValidations.baseValidation;
 import static io.github.epam.states.States.shouldBeLoggedIn;
-import static io.github.epam.test.data.InputData.inputDataAsMap;
+import static io.github.epam.test.data.InputDataHelper.inputDataAsMap;
 import static org.hamcrest.Matchers.containsString;
 import static org.testng.Assert.assertEquals;
 import static org.testng.Assert.assertTrue;
