+++ jdi-light-bootstrap-tests/src/test/java/io/github/epam/test/data/InputDataHelper.java
@@ -1,11 +1,9 @@
 package io.github.epam.test.data;
 
-import java.util.ArrayList;
 import java.util.HashMap;
-import java.util.List;
 import java.util.Map;
 
-public class InputData {
+public class InputDataHelper {
     public static Map<Integer, String> inputDataAsMap(int size) {
         Map<Integer, String> inputData = new HashMap<Integer, String>(size);
 
@@ -15,6 +13,4 @@ public class InputData {
 
         return inputData;
     }
-
-    public static String TEXT = "data";
 }
